Various fixes (#381)
* fix fixture instance type

* update logic in waiting for disabled distribution

* add test for restart_stalled_pipelines

* update disassociate_health_check to handle instance with missing columns

* update delete_health_checks to handle missing columns

* add log line to code

* add test for delete_web_acl to handle instances with missing columns

* refactor code

* update delete_health_check_alarms to only delete if there are alarms to delete

+++ broker/tasks/route53.py
@@ -328,8 +328,14 @@ def delete_health_checks(operation_id: int, **kwargs):
 
     logger.info(f'Deleting health check(s) for "{service_instance.domain_names}"')
 
+    existing_health_checks = service_instance.route53_health_checks
+
+    if existing_health_checks == None:
+        logger.info("No Route53 health checks to delete")
+        return
+
     updated_health_checks = _delete_health_checks(
-        service_instance.route53_health_checks, service_instance.route53_health_checks
+        existing_health_checks, existing_health_checks
     )
 
     service_instance.route53_health_checks = updated_health_checks
